+++ src/main/java/dk/kb/storage/storage/DsStorage.java
@@ -14,6 +14,7 @@ import dk.kb.storage.model.v1.MappingDto;
 import dk.kb.storage.model.v1.OriginCountDto;
 import dk.kb.storage.model.v1.RecordTypeDto;
 import dk.kb.storage.model.v1.RecordsCountDto;
+import dk.kb.storage.model.v1.TranscriptionDto;
 import dk.kb.storage.util.UniqueTimestampGenerator;
 
 
@@ -39,10 +40,11 @@ public class DsStorage implements AutoCloseable {
     private static SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ssZ",Locale.getDefault());
     
     
+    private static final String TRANSCRIPTIONS_TABLE = "ds_transcriptions";
     private static final String RECORDS_TABLE = "ds_records";
     private static final String MAPPING_TABLE = "ds_mapping";
-    private static final String ID_COLUMN = "id";
     private static final String ORGID_COLUMN = "orgid";
+    private static final String ID_COLUMN = "id";
     private static final String IDERROR_COLUMN = "id_error";
     private static final String ORIGIN_COLUMN = "origin";
     private static final String RECORDTYPE_COLUMN = "recordtype";
@@ -55,7 +57,12 @@ public class DsStorage implements AutoCloseable {
     private static final String RECORDS_KALTURA_ID_COLUMN = "kalturaid";
     private static final String MAPPING_REFERENCE_ID_COLUMN = "referenceid";
     private static final String MAPPING_KALTURA_ID_COLUMN = "kalturaid";
-
+       
+    private static final String FILE_ID_COLUMN = "fileid";
+    private static final String FILE_NAME_COLUMN = "filename";
+    private static final String TRANSCRIPTION_TEXT_COLUMN= "transcription";
+    private static final String TRANSCRIPTION_LINES_COLUMN = "transcription_lines";
+    
     private static String createRecordStatement = "INSERT INTO " + RECORDS_TABLE +
             " (" + ID_COLUMN + ", " + ORIGIN_COLUMN + ", " +ORGID_COLUMN + ","+ RECORDTYPE_COLUMN +"," + IDERROR_COLUMN +","+ DELETED_COLUMN + ", " + CTIME_COLUMN + ", " + MTIME_COLUMN + ", " + DATA_COLUMN + ", " + PARENT_ID_COLUMN +  " , " + RECORDS_REFERENCE_ID_COLUMN +" , "+RECORDS_KALTURA_ID_COLUMN+")"+
             " VALUES (?,?,?,?,?,?,?,?,?,?,?,?)";
@@ -122,6 +129,11 @@ public class DsStorage implements AutoCloseable {
 
     private static String recordByIdStatement = "SELECT * FROM " + RECORDS_TABLE + " WHERE ID= ?";
 
+    
+    private static String transcriptionByFileIdStatement = "SELECT * FROM " + TRANSCRIPTIONS_TABLE+ " WHERE "+FILE_ID_COLUMN +" = ?";
+    
+    private static String transcriptionByFileIdCountStatement = "SELECT count(*) as count FROM " + TRANSCRIPTIONS_TABLE+ " WHERE "+FILE_ID_COLUMN +" = ?";
+    
     // SELECT mtime FROM ds_records WHERE origin= 'test_base' ORDER BY mtime DESC
     private static final String maxMtimeStatement =
             "SELECT " + MTIME_COLUMN + " FROM " + RECORDS_TABLE +
@@ -211,6 +223,18 @@ public class DsStorage implements AutoCloseable {
             " ORDER BY "+MTIME_COLUMN+ " ASC LIMIT ?";
 
 
+    //SELECT * FROM  ds_records  WHERE origin= 'test_origin' AND mtime  > 1637237120476001 AND parentId IS NULL ORDER BY mtime ASC LIMIT 100    
+    private static String createTranscriptionStatement =
+            "INSERT INTO " + TRANSCRIPTIONS_TABLE +
+            " (" +
+              FILE_ID_COLUMN + ", "+
+              FILE_NAME_COLUMN +", "+
+              MTIME_COLUMN +", "+
+              TRANSCRIPTION_TEXT_COLUMN+ ", "+
+              TRANSCRIPTION_LINES_COLUMN+") "+
+            " VALUES (?,?,?,?,?)";
+
+    
     //Optimized SQL that finds missing KalturaIds on records table that have a referenceId but no kalturaId. Can take some time(minutes) first time if millions of records miss
     // kalturaId
     //'INNER JOIN' will only return matched rows from both table  unlike 'LEFT JOIN' that will return non-matched also. 
@@ -222,6 +246,8 @@ public class DsStorage implements AutoCloseable {
                                                         " WHERE  A."+RECORDS_KALTURA_ID_COLUMN+" IS NULL AND b."+MAPPING_KALTURA_ID_COLUMN +" IS NOT NULL";    
     
     private static String originsStatisticsStatement = "SELECT " + ORIGIN_COLUMN + " ,COUNT(*) AS COUNT , SUM("+DELETED_COLUMN+") AS deleted,  max("+MTIME_COLUMN + ") AS MAX FROM " + RECORDS_TABLE + " group by " + ORIGIN_COLUMN;
+
+    private static String deleteTranscriptionByFileIdStatement = "DELETE FROM " + TRANSCRIPTIONS_TABLE + " WHERE "+FILE_ID_COLUMN+" = ?";
     private static String deleteMarkedForDeleteStatement = "DELETE FROM " + RECORDS_TABLE + " WHERE "+ORIGIN_COLUMN +" = ? AND "+DELETED_COLUMN +" = 1" ;   
     private static String recordIdExistsStatement = "SELECT COUNT(*) AS COUNT FROM " + RECORDS_TABLE+ " WHERE "+ID_COLUMN +" = ?";
     private static String countRecordsInOriginStatement = "SELECT COUNT(*) FROM " + RECORDS_TABLE +  " WHERE " + ORIGIN_COLUMN + " = ? AND " + MTIME_COLUMN + " > ?";
@@ -751,6 +777,30 @@ public class DsStorage implements AutoCloseable {
 
     }
 
+    public void createNewTranscription(TranscriptionDto transcription) throws Exception {
+
+        // Sanity check
+        if (transcription.getFileId() == null) {
+            throw new Exception("Fileid must not be null");
+        }
+        long nowStamp = UniqueTimestampGenerator.next();
+
+        try (PreparedStatement stmt = connection.prepareStatement(createTranscriptionStatement)) {
+            stmt.setString(1, transcription.getFileId());
+            stmt.setString(2, transcription.getFileName());                        
+            stmt.setLong(3, nowStamp);            
+            stmt.setString(4, transcription.getTranscription());           
+            stmt.setString(5, transcription.getTranscriptionLines());
+            stmt.executeUpdate();
+        } catch (SQLException e) {
+            String message = "SQL Exception in createNewTranscription with fileid:" + transcription.getFileId() + " error:" + e.getMessage();
+            log.error(message);
+            throw new SQLException(message, e);
+        }
+    }
+        
+    
+    
     /**
      * Update the modified time for input record.
      * @param recordId of record to update
@@ -953,7 +1003,26 @@ public class DsStorage implements AutoCloseable {
 
     }
 
-    
+    /**
+     * 
+     * Delete a transcription by fileid
+     * 
+     * @param fileId the fileId. If fileId is not found nothing will be deleted, but it will be logged.
+     * @throws Exception Only if unexpected SQL exception happens.
+     */     
+    public void deleteTranscriptionByFileId(String fileId) throws Exception {    
+        try (PreparedStatement stmt = connection.prepareStatement(deleteTranscriptionByFileIdStatement)) {        
+            stmt.setString(1, fileId);
+            int numberDeleted = stmt.executeUpdate();            
+            if (numberDeleted != 1) {
+                log.warn("Delete transcription by fileId did not delete 1 as expected. Deleted='{}', FileId='{}'",numberDeleted,fileId);
+            }            
+        } catch (SQLException e) {
+            String message = "SQL Exception in deleteTranscriptionByFileId for fileId:" + fileId + " error:" + e.getMessage();
+            log.error(message);
+            throw new SQLException(message, e);
+        }
+    }
     /**
      * 
      * Update a mapping with a new kalturaId
@@ -1112,7 +1181,58 @@ public class DsStorage implements AutoCloseable {
         return record;
     }
 
+    /**
+     * Load a transcription by fileId.
+     * 
+     * @param fileId the fileId to load
+     * @return TranscriptionDto. If fileId is not found will return null
+     */
+    public TranscriptionDto loadTranscription(String fileId) throws SQLException {
+        try (PreparedStatement stmt = connection.prepareStatement(transcriptionByFileIdStatement )) {
+            stmt.setString(1, fileId);
+
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (!rs.next()) {
+                    return null;
+                }
+                TranscriptionDto trans = createTranscriptionFromRS(rs);                            
+                return trans;
+            }        
+        }
+    }
+        
+    /**
+     * Count number of transcriptions by fileId. This is a fast method so see if a transcriptions exists instead of loading all text.
+     * 
+     * @param fileId the fileId count
+     * @return 0 or 1. FileId is unique
+     */
+    public int countTranscriptionByFileId(String fileId) throws SQLException {
+        try (PreparedStatement stmt = connection.prepareStatement( transcriptionByFileIdCountStatement)) {
+            stmt.setString(1, fileId);
+            try (ResultSet rs = stmt.executeQuery()) { 
+                rs.next();//always value                
+                return rs.getInt("count");            
+            }        
+        }
+    }
     
+    private static TranscriptionDto createTranscriptionFromRS(ResultSet rs) throws SQLException {
+        String fileId = rs.getString(FILE_ID_COLUMN);
+        String fileName= rs.getString(FILE_NAME_COLUMN);         
+        long mTime = rs.getLong(MTIME_COLUMN);
+        String transcriptionText = rs.getString(TRANSCRIPTION_TEXT_COLUMN);
+        String transcriptionLines = rs.getString(TRANSCRIPTION_LINES_COLUMN);
+        
+        TranscriptionDto transcription = new TranscriptionDto();
+        transcription.setFileId(fileId);
+        transcription.setFileName(fileName);
+        transcription.setmTime(mTime);
+        transcription.setTranscription(transcriptionText);
+        transcription.setTranscriptionLines(transcriptionLines);
+        return transcription;
+    }
+
     private static int boolToInt(Boolean isTrue) {
         if (isTrue == null) {
             return 0;