+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- New table in database (ds_transcriptions). DDL to create the table must be run for new release.
+- New service method transcription(POST) to add or update a transcription
+- New service method to load a transcription. Key is the external fileId (filename)
 
 ### Changed